Fixing formatting bugs

+++ plugins/plugins/XactThrashingScenario.java
@@ -1400,8 +1400,8 @@ if(_clientNum % 100 == 0){
 		int batchSetID = stepA(transactionSize, eXclusiveLocks, effectiveDBSize);
 		Main._logger.outputLog(String.format(
 				"Start the batchSet #%d(runID:%d,xactSz:%.2f%%/xLcks:%d%%/effDBSz:%d%%) analysis!",
-				batchSetID, runID, transactionSize*100, eXclusiveLocks*100,
-				effectiveDBSize*100));
+				batchSetID, runID, transactionSize*100, (int)(eXclusiveLocks*100),
+				(int)(effectiveDBSize*100)));
 
 		// make a batchset run result
 		int batchSetRunResID = insertBatchSetRunResult(runID, batchSetID);
